requirements.txt
alembic==1.8.0
anyio==3.6.1
astroid==2.11.7
attrs==21.4.0
bcrypt==3.2.2
black==22.6.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
decouple==0.0.7
dill==0.3.5.1
distlib==0.3.4
factory-boy==3.2.1
Faker==13.15.0
fastapi==0.78.0
filelock==3.7.1
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==1.7.0
importlib-resources==5.8.0
iniconfig==1.1.1
isort==5.10.1
keyring==22.3.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==20.9
passlib==1.7.4
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.1.14
poetry-core==1.0.8
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pylev==1.4.0
pylint==2.14.4
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
virtualenv==20.15.1
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.0
